Zmiany

+++ main.cpp
@@ -13,7 +13,7 @@ struct Node {
 };
 
 //przejście wszerz z ograniczeniem wysokości
-bool bfs(vector<Node>& arr, int start, int end, int water_level) {
+bool bfs(vector<Node>& arr, int start, int end, long water_level) {
     if (arr[start].height < water_level || arr[end].height < water_level) {
         return false;
     }
@@ -28,7 +28,6 @@ bool bfs(vector<Node>& arr, int start, int end, int water_level) {
 
     while (!q.empty()) {
         int z = q.front();
-        //std::cout << z << "\n";
         q.pop();
 
         for (int neighbor : arr[z].neighbors) {
@@ -60,39 +59,35 @@ int main() {
 
     for (int i = 0; i < szczyty; ++i) {
         std::cin >> graph[i].height;
-        //std::cout << "Wierzchołek " << i << " ma wysokość " << graph[i].height << "\n";
     }
 
     for (int i = 0; i < krawedzie; i++) {
         int a, b;
         std::cin >> a;
-        //std::cout << "a: " << a << "\n";
 
         // dodawanie sąsiada tylko jeśli jest podany
         if (std::cin >> b) {
             graph[a].neighbors.push_back(b);
+            //graph[b].neighbors.push_back(a);
         }
-        //std::cout << "b: " << b << "\n";
     }
 
     // wczytywanie punktu początkowego i końcowego
     int start, end;
     std::cin >> start >> end;
-    //std::cout << "Start: " << start << ", End: " << end << "\n";
 
     // wczytywanie zmian poziomu wody
-    int current_water_level = 0;
+    long current_water_level = 0;
     short zmiany_poziomu_wody;
     std::cin >> zmiany_poziomu_wody;
 
     // do przechowywania wyników dla różnych poziomów wody (może zawierać wartości ujemne)
-    std::unordered_map<int, string> result_cache;
+    std::unordered_map<long, string> result_cache;
 
     for (int i = 0; i < zmiany_poziomu_wody; i++) {
         int change;
         std::cin >> change;
         current_water_level += change;
-        //std::cout << "Aktualny poziom wody: " << current_water_level << "\n";
 
         if (result_cache[current_water_level].empty()) {
             string result = bfs(graph, start, end, current_water_level) ? "TAK\n" : "NIE\n";
